# HDDS-5793. Speed up TestBlockOutputStreamWithFailures by combining test cases