--- a/hadoop-ozone/integration-test/src/test/java/org/apache/hadoop/ozone/client/rpc/TestBlockOutputStreamWithFailuresFlushDelay.java
+++ b/hadoop-ozone/integration-test/src/test/java/org/apache/hadoop/ozone/client/rpc/TestBlockOutputStreamWithFailuresFlushDelay.java
@@ -162,6 +162,21 @@ public class TestBlockOutputStreamWithFailuresFlushDelay {
   }
 
   @Test
+  public void testThreeNode() throws Exception {
+    testWatchForCommitWithCloseContainerException();
+    testWatchForCommitDatanodeFailure();
+    test2DatanodesFailure();
+    testFailureWithPrimeSizedData();
+    testExceptionDuringClose();
+  }
+
+  @Test
+  public void testOneNode() throws Exception {
+    testDatanodeFailureWithSingleNodeRatis();
+    testWatchForCommitWithSingleNodeRatis();
+    testDatanodeFailureWithPreAllocation();
+  }
+
   public void testWatchForCommitWithCloseContainerException()
       throws Exception {
     String keyName = getKeyName();
@@ -251,7 +266,6 @@ public class TestBlockOutputStreamWithFailuresFlushDelay {
     validateData(keyName, dataString.concat(dataString).getBytes(UTF_8));
   }
 
-  @Test
   public void testWatchForCommitDatanodeFailure() throws Exception {
     String keyName = getKeyName();
     OzoneOutputStream key = createKey(keyName, ReplicationType.RATIS, 0);
@@ -337,8 +351,7 @@ public class TestBlockOutputStreamWithFailuresFlushDelay {
     String dataString = new String(data1, UTF_8);
     validateData(keyName, dataString.concat(dataString).getBytes(UTF_8));
   }
-  
-  @Test
+
   public void test2DatanodesFailure() throws Exception {
     String keyName = getKeyName();
     OzoneOutputStream key = createKey(keyName, ReplicationType.RATIS, 0);
@@ -443,7 +456,6 @@ public class TestBlockOutputStreamWithFailuresFlushDelay {
     validateData(keyName, data1);
   }
 
-  @Test
   public void testFailureWithPrimeSizedData() throws Exception {
     String keyName = getKeyName();
     OzoneOutputStream key = createKey(keyName, ReplicationType.RATIS, 0);
@@ -507,7 +519,6 @@ public class TestBlockOutputStreamWithFailuresFlushDelay {
     validateData(keyName, dataString.concat(dataString).getBytes(UTF_8));
   }
 
-  @Test
   public void testExceptionDuringClose() throws Exception {
     String keyName = getKeyName();
     OzoneOutputStream key = createKey(keyName, ReplicationType.RATIS, 0);
@@ -578,7 +589,6 @@ public class TestBlockOutputStreamWithFailuresFlushDelay {
     validateData(keyName, dataString.concat(dataString).getBytes(UTF_8));
   }
 
-  @Test
   public void testWatchForCommitWithSingleNodeRatis() throws Exception {
     String keyName = getKeyName();
     OzoneOutputStream key =
@@ -669,7 +679,6 @@ public class TestBlockOutputStreamWithFailuresFlushDelay {
     validateData(keyName, dataString.concat(dataString).getBytes(UTF_8));
   }
 
-  @Test
   public void testDatanodeFailureWithSingleNodeRatis() throws Exception {
     String keyName = getKeyName();
     OzoneOutputStream key =
@@ -762,7 +771,6 @@ public class TestBlockOutputStreamWithFailuresFlushDelay {
     validateData(keyName, dataString.concat(dataString).getBytes(UTF_8));
   }
 
-  @Test
   public void testDatanodeFailureWithPreAllocation() throws Exception {
     String keyName = getKeyName();
     OzoneOutputStream key =