## What changes were proposed in this pull request?

Combine some test cases in `TestBlockOutputStreamWithFailures` and `TestBlockOutputStreamWithFailuresFlushDelay` to reduce cluster creation/startup time.

https://issues.apache.org/jira/browse/HDDS-5793

## How was this patch tested?

Before:

```
[INFO] Tests run: 8, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 585.121 s - in org.apache.hadoop.ozone.client.rpc.TestBlockOutputStreamWithFailuresFlushDelay
[INFO] Tests run: 8, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 584.137 s - in org.apache.hadoop.ozone.client.rpc.TestBlockOutputStreamWithFailures
```

https://github.com/apache/ozone/runs/3728055559#step:4:2863

After:

```
[INFO] Tests run: 2, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 294.082 s - in org.apache.hadoop.ozone.client.rpc.TestBlockOutputStreamWithFailuresFlushDelay
[INFO] Tests run: 2, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 278.229 s - in org.apache.hadoop.ozone.client.rpc.TestBlockOutputStreamWithFailures
```

https://github.com/adoroszlai/hadoop-ozone/runs/3732932406#step:4:2892