@adoroszlai 2021-09-29
Slightly less time saved with the new grouping, but hopefully it is more stable.